# api/openapi/vela.yaml
openapi: 3.0.3
info:
  title: Vela API
  version: 1.0.0
  description: Project-scoped asynchronous AI generation API.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /v1/projects/{project_id}/jobs:
    post:
      operationId: submitJob
      summary: Submit a Job for durable Admission
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '202':
          description: Job accepted durably
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '402':
          description: Contract Credit Limit is insufficient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Idempotency key was accepted with different content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Project queue limit is exhausted
          headers:
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Compatible capacity is unavailable
          headers:
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/jobs/{job_id}:
    get:
      operationId: getJob
      summary: Get authoritative Project Job state
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Current Job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Job is not visible in this Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/jobs/{job_id}/cancel:
    post:
      operationId: cancelJob
      summary: Commit a Customer Cancellation decision
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Cancellation committed or replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Job is not visible in this Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{project_id}/jobs/{job_id}/artifacts:
    get:
      operationId: getJobArtifacts
      summary: Get the committed ArtifactSet with short-lived exact-version URLs
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Committed ArtifactSet and exact-version downloads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactSet'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: ArtifactSet is not visible or not currently accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VelaCredential
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
  headers:
    RetryAfter:
      schema:
        type: integer
        minimum: 1
      description: Seconds before the client should retry Admission.
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Credential is absent or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Principal lacks the required Project scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    SubmitJobRequest:
      type: object
      additionalProperties: false
      required:
        - model
        - generation_preset
        - service_class
        - output_spec
        - generation_count
        - prompt
      properties:
        model:
          type: string
          minLength: 1
          maxLength: 100
        generation_preset:
          type: string
          enum:
            - quality
            - balanced
            - fast
        service_class:
          type: string
          enum:
            - standard
        output_spec:
          type: string
          minLength: 1
          maxLength: 100
        generation_count:
          type: integer
          minimum: 1
          maximum: 16
        prompt:
          type: string
          minLength: 1
          maxLength: 20000
        client_metadata:
          type: object
          additionalProperties: true
          x-go-type: json.RawMessage
    Job:
      type: object
      additionalProperties: false
      required:
        - job_id
        - project_id
        - state
        - attempts_started
        - pricing
        - job_expires_at
        - created_at
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/JobState'
        phase:
          $ref: '#/components/schemas/ExecutionPhase'
        phase_progress:
          type: number
          format: double
          minimum: 0
          maximum: 1
          exclusiveMaximum: true
        attempts_started:
          type: integer
          minimum: 0
        next_retry_at:
          type: string
          format: date-time
        estimated_finish_at:
          type: string
          format: date-time
        progress_updated_at:
          type: string
          format: date-time
        pricing:
          $ref: '#/components/schemas/PricingSnapshot'
        job_expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    JobState:
      type: string
      enum:
        - QUEUED
        - ASSIGNED
        - RUNNING
        - FINALIZING
        - RETRY_WAIT
        - CANCELING
        - SUCCEEDED
        - FAILED
        - CANCELED
    CancelResult:
      type: object
      additionalProperties: false
      required:
        - cancellation_id
        - job_id
        - decision
        - state
        - job_version
        - billable
        - decided_at
      properties:
        cancellation_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/CancelDecision'
        state:
          $ref: '#/components/schemas/JobState'
        job_version:
          type: integer
          format: int64
          minimum: 1
        billable:
          type: boolean
        charge:
          $ref: '#/components/schemas/Charge'
        decided_at:
          type: string
          format: date-time
    CancelDecision:
      type: string
      enum:
        - CANCELED
        - CANCELING
        - ALREADY_SUCCEEDED
        - ALREADY_FAILED
    Charge:
      type: object
      additionalProperties: false
      required:
        - charge_id
        - amount_minor
        - currency
        - reason
        - posted_at
      properties:
        charge_id:
          type: string
          format: uuid
        amount_minor:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        reason:
          type: string
          enum:
            - VISIBLE_COMPLETION
            - CUSTOMER_CANCELLATION
        posted_at:
          type: string
          format: date-time
    ArtifactSet:
      type: object
      additionalProperties: false
      required:
        - artifact_set_id
        - job_id
        - retention_expires_at
        - committed_at
        - artifacts
      properties:
        artifact_set_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        retention_expires_at:
          type: string
          format: date-time
        committed_at:
          type: string
          format: date-time
        artifacts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ArtifactDownload'
    ArtifactDownload:
      type: object
      additionalProperties: false
      required:
        - artifact_id
        - kind
        - ordinal
        - object_version_id
        - size_bytes
        - sha256
        - content_type
        - download_url
        - download_url_expires_at
      properties:
        artifact_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - VIDEO
            - THUMBNAIL
        ordinal:
          type: integer
          format: int32
          minimum: 0
        object_version_id:
          type: string
          minLength: 1
          maxLength: 1000
        size_bytes:
          type: integer
          format: int64
          minimum: 1
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        content_type:
          type: string
          minLength: 1
          maxLength: 200
        download_url:
          type: string
          format: uri
        download_url_expires_at:
          type: string
          format: date-time
    ExecutionPhase:
      type: string
      enum:
        - QUEUED
        - PREPARING
        - GENERATING
        - FINALIZING
        - RETRY_WAIT
    PricingSnapshot:
      type: object
      additionalProperties: false
      required:
        - rate_card_revision_id
        - rate_line_id
        - unit_amount_minor
        - quantity
        - quoted_amount_minor
        - currency
      properties:
        rate_card_revision_id:
          type: string
          format: uuid
        rate_line_id:
          type: string
          format: uuid
        unit_amount_minor:
          type: integer
          format: int64
          minimum: 0
        quantity:
          type: integer
          minimum: 1
        quoted_amount_minor:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string